pyproject.toml
[tool.poetry]
name = "deep-learning-package"
version = "0.1.0"
description = ""
authors = ["Ko-nifuramu <kotaro22235963@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.4"
torch = "^2.2.0"
matplotlib = "^3.8.3"
tqdm = "^4.66.2"
pyyaml = "^6.0.1"
scikit-learn = "^1.4.1.post1"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
pre-commit = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
